requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dill==0.4.0
docling==2.40.0
docling-core==2.40.0
docling-ibm-models==3.7.0
docling-parse==4.1.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
latex2mathml==3.78.0
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pypandoc_binary==1.15
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
semchunk==2.2.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5
